requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
beautifulsoup4==4.14.3
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
colorlog==6.10.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.13
decorator==5.2.1
distro==1.9.0
einops==0.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.1
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.22.2
HyperPyYAML==1.2.3
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
julius==0.2.7
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
lightning==2.6.0
lightning-utilities==0.15.2
llvmlite==0.46.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.15
networkx==3.4.2
nltk==3.9.2
noisereduce==3.0.3
numba==0.63.1
numpy==1.26.4
omegaconf==2.3.0
openai==2.14.0
openai-whisper==20250625
optuna==4.6.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.5.1
pooch==1.8.2
preshed==3.0.12
primePy==1.3
propcache==0.4.1
protobuf==6.33.2
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannote.audio==3.1.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.1.3
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
scikit-learn==1.7.2
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.8.1
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.3
SQLAlchemy==2.0.45
srsly==2.5.2
sympy==1.14.0
tensorboardX==2.6.4
textblob==0.19.0
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tomli==2.3.0
torch==2.1.2+cu118
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.1.2+cu118
torchmetrics==1.8.2
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
yarl==1.22.0
